refactor: extract message rendering into specMessages.ts, fix fallback text
messages2Text/getMessageString move out as free functions (class keeps
static delegates for the npm CLI contract). Fixes the getMessageLocal
bug: messages without referencedEntity/additionalInformation now return
their actual text instead of a hardcoded 'This will break compatibilty'
warning (visible e.g. in contribute error messages); the never-passed
notBackwardCompatible parameter and the pointless structuredClone of a
string are gone.
