+++ backend/src/specification/index.ts
@@ -4,6 +4,7 @@ export * from './log.js'
 export * from './ifilespecification.js'
 export * from './m2mspecification.js'
 export * from './modbusValues.js'
+export * from './specMessages.js'
 export * from './m2mgithub.js'
 export * from './configspec.js'
 export * from './migrator.js'
